pyproject.toml
[project]
name = "immichporter"
version = "0.0.2"
description = "Import photos from various sources to Immich"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Multimedia :: Graphics :: Graphics Conversion",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
  "playwright",
  "rich",
  "python-dateutil",
  "click",
  "sqlalchemy",
  "loguru",
  "platformdirs>=4.4.0",
  "prompt-toolkit>=3.0.52",
  "attrs>=25.3.0",
  "httpx>=0.28.1",
  "pydantic>=2.11.10",
  "psycopg2-binary>=2.9.11",
]

[project.scripts]
immichporter = "immichporter.cli:main"
imp = "immichporter.cli:main"

[project.urls]
Repository = "https://github.com/burgdev/immichporter"
Documentation = "https://burgdev.github.io/immichporter/"
Changelog = "https://github.com/burgdev/immichporter/blob/main/CHANGELOG.md"
Releases = "https://github.com/burgdev/immichporter/releases"

[dependency-groups]
dev = [
  "bump2version>=1.0.1",
  "deptry>=0.21.1",
  "environs>=11.2.1",
  "git-cliff>=2.7.0",
  "humanize>=4.11.0",
  "invoke>=2.2.0",
  "jinja2>=3.1.6",
  "loguru>=0.7.3",
  "openapi-python-client>=0.26.1",
  "pillow>=11.3.0",
  "pre-commit>=4.0.1",
  "pyright>=1.1.0",
  "rich>=13.9.4",
  "ruff>=0.8.2",
  "toml>=0.10.2",
  "types-requests>=2.28.0",
]

docs = [
  "mkdocs>=1.4.0",
  "termynal>=0.12.2",
  "mkdocs-autoapi>=0.3.2",
  "mkdocs-click>=0.8.1",
  "mkdocs-gen-files>=0.5.0",
  "mkdocs-literate-nav>=0.6.1",
  "mkdocs-material>=9.5.49",
  "mkdocs-open-in-new-tab>=1.0.8",
  "mkdocs-section-index>=0.3.9",
  "mkdocstrings[python]>=0.22.0",
]

test = [
  "pytest>=8.3.4",
  "pytest-cov>=4.1.0",
  "pytest-docker>=1.0.0",
  "pytest-xdist>=3.8.0",
]


[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-third-party = ["immichporter"]

[tool.deptry]
extend_exclude = ["tasks/__init__.py"]

[tool.deptry.per_rule_ignores]
DEP004 = [
  "toml",
  "rich",
  "invoke",
  "environs",
  "mkdocs_gen_files",
  "jinja2",
  "loguru",
] # only in dev
DEP003 = ["immichporter"] # transitive dependencies
DEP002 = ["psycopg2-binary"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/immichporter"]

[tool.hatch.build.targets.sdist]
include = ["src/immichporter"]
